# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/register": {
      "post": {
        "operationId": "UsersController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "用户注册",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "用户登录",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "获取单个用户",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "更新用户",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        },
        "summary": "删除用户",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/transactions": {
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        },
        "summary": "创建交易记录",
        "tags": [
          "Transactions"
        ]
      },
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取所有交易记录",
        "tags": [
          "Transactions"
        ]
      }
    },
...
    "/api/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "交易记录ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        },
        "summary": "获取单个交易记录",
        "tags": [
          "Transactions"
        ]
      },
      "put": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "交易记录ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        },
        "summary": "更新交易记录",
        "tags": [
          "Transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "交易记录ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        },
        "summary": "删除交易记录",
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "small-ledger API",
    "description": "家有小账本服务端API文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "minLength": 3,
            "maxLength": 50
          },
          "password": {
            "type": "string",
            "description": "密码",
            "minLength": 6,
            "maxLength": 255
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "username",
          "created_at",
          "updated_at"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "minLength": 3,
            "maxLength": 50
          },
          "password": {
            "type": "string",
            "description": "密码",
            "minLength": 6,
            "maxLength": 255
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "billId": {
            "type": "string",
            "description": "账单ID"
          },
          "platform": {
            "type": "string",
            "description": "支付平台",
            "enum": [
              "alipay",
              "wechat_pay"
            ]
          },
          "shop": {
            "type": "string",
            "description": "商店名称"
          },
          "product": {
            "type": "string",
            "description": "产品名称"
          },
          "type": {
            "type": "string",
            "description": "交易类型",
            "enum": [
              "income",
              "expense",
              "neutral"
            ]
          },
          "amount": {
            "type": "number",
            "description": "交易金额",
            "minimum": 0.01
          },
          "category": {
            "type": "string",
            "description": "交易分类"
          },
          "description": {
            "type": "string",
            "description": "交易描述"
          },
          "transactionDate": {
            "format": "date-time",
            "type": "string",
            "description": "交易日期"
          },
          "transactionStartDate": {
            "format": "date-time",
            "type": "string",
            "description": "交易开始日期"
          },
          "transactionEndDate": {
            "format": "date-time",
            "type": "string",
            "description": "交易结束日期"
          }
        },
        "required": [
          "billId",
          "platform",
          "shop",
          "product",
          "type",
          "amount",
          "category",
          "transactionStartDate",
          "transactionEndDate"
        ]
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "交易记录ID"
          },
          "bill_id": {
            "type": "string",
            "description": "账单ID"
          },
          "platform": {
            "type": "string",
            "description": "支付平台",
            "enum": [
              "alipay",
              "wechat_pay"
            ]
          },
          "shop": {
            "type": "string",
            "description": "店铺名称"
          },
          "product": {
            "type": "string",
            "description": "产品名称"
          },
          "type": {
            "type": "string",
            "description": "交易类型",
            "enum": [
              "income",
              "expense",
              "neutral"
            ]
          },
          "amount": {
            "type": "number",
            "description": "交易金额"
          },
          "category": {
            "type": "string",
            "description": "交易分类"
          },
          "description": {
            "type": "string",
            "description": "交易描述"
          },
          "transaction_date": {
            "format": "date-time",
            "type": "string",
            "description": "交易日期"
          },
          "transaction_start_date": {
            "format": "date-time",
            "type": "string",
            "description": "交易开始日期"
          },
          "transaction_end_date": {
            "format": "date-time",
            "type": "string",
            "description": "交易结束日期"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "bill_id",
          "platform",
          "shop",
          "product",
          "type",
          "amount",
          "category",
          "description",
          "transaction_date",
          "transaction_start_date",
          "transaction_end_date",
          "created_at",
          "updated_at"
        ]
      },
...
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}